+++ CHANGELOG.md
@@ -137,4 +137,5 @@
 ### v1.0 8/11/19
 * debugged success message in file copy
 * added success and failure events to os check
-
+* reformatted domain_sticate.sh to python and added lib_install.sh check and shodan search to domain scan
+* added domain_sticate.py to menu